reimplementation of :terminal to set proper width
Problem:
Removing number column etc in TermOpen is too late for `:terminal cmd`.
In f_termopen(), channel_job_start() executes the cmd with the current
window width, and TermOpen is triggered after that. This affects styling
of git-delta run with :T.

Solution:
Custom implementation of :terminal that removes the columns before
invoking termopen().